Use decimals and fix invoices

+++ BTCPayServer/Services/Invoices/InvoiceEntity.cs
@@ -358,9 +358,9 @@ namespace BTCPayServer.Services.Invoices
                 cryptoInfo.CryptoPaid = accounting.CryptoPaid.ToString();
 
                 cryptoInfo.Address = info.GetPaymentMethodDetails()?.GetPaymentDestination();
-                cryptoInfo.ExRates = new Dictionary<string, double>
+                cryptoInfo.ExRates = new Dictionary<string, decimal>
                 {
-                    { ProductInformation.Currency, (double)cryptoInfo.Rate }
+                    { ProductInformation.Currency, cryptoInfo.Rate }
                 };
                 var paymentId = info.GetId();
                 var scheme = info.Network.UriScheme;
